+++ dinghy-lib/src/apple/device.rs
@@ -57,7 +57,12 @@ impl IosDevice {
     }
 
     fn is_pre_ios_17(&self) -> Result<bool> {
-        Ok(semver::VersionReq::parse(&self.os)?.comparators.get(0).ok_or_else(|| anyhow!("Invalid iOS version: {}", self.os))?.major < 17)
+        Ok(semver::VersionReq::parse(&self.os)?
+            .comparators
+            .get(0)
+            .ok_or_else(|| anyhow!("Invalid iOS version: {}", self.os))?
+            .major
+            < 17)
     }
 
     fn is_locked(&self) -> Result<bool> {
@@ -547,7 +552,14 @@ fn launch_app(dev: &AppleSimDevice, app_args: &[&str], _envs: &[&str]) -> Result
         .to_string_lossy()
         .into_owned();
     let stdout_param = &format!("--stdout={}", stdout);
-    let mut xcrun_args: Vec<&str> = vec!["simctl", "launch", "-w", stdout_param, &dev.id, "Dinghy"];
+    let mut xcrun_args: Vec<&str> = vec![
+        "simctl",
+        "launch",
+        "--wait-for-debugger",
+        stdout_param,
+        &dev.id,
+        "Dinghy",
+    ];
     xcrun_args.extend(app_args);
     debug!("Launching app via xcrun using args: {:?}", xcrun_args);
     let launch_output = process::Command::new("xcrun")
@@ -572,7 +584,8 @@ fn launch_app(dev: &AppleSimDevice, app_args: &[&str], _envs: &[&str]) -> Result
         .arg("-s")
         .arg(lldb_script_filename)
         .output()?;
-    let test_contents = std::fs::read_to_string(stdout)?;
+    let test_contents = std::fs::read_to_string(&stdout)
+        .with_context(|| format!("Reading llvm stdout from {stdout}"))?;
     println!("{}", test_contents);
 
     let output: String = String::from_utf8_lossy(&output.stdout).to_string();