+++ dinghy-lib/src/apple/mod.rs
@@ -19,11 +19,13 @@ pub struct SignatureSettings {
     pub file: String,
     pub entitlements: String,
     pub name: String,
+    #[allow(dead_code)]
     pub profile: String,
 }
 
 #[derive(Debug, Clone)]
 pub struct SigningIdentity {
+    #[allow(dead_code)]
     pub id: String,
     pub name: String,
     pub team: String,
@@ -319,11 +321,15 @@ fn devices_from_devicectl(devices: &mut HashMap<String, IosDevice>) -> Result<()
     if !devicectl.status.success() {
         bail!("xcrun command failed. Please check that \"xcrun devicectl list devices\" works.\n{devicectl:?}");
     }
-    for device in json::parse(&std::fs::read_to_string(tmpjson)?)?["result"]["devices"].members() {
-        let udid = device["hardwareProperties"]["udid"]
+    let txt = std::fs::read_to_string(&tmpjson)
+        .with_context(|| format!("Reading devicectl json output {tmpjson:?}"))?;
+    for device in json::parse(&txt)?["result"]["devices"].members() {
+        let Some(udid) = device["hardwareProperties"]["udid"]
             .as_str()
-            .context("no identifier in device json")?
-            .to_string();
+            .map(|s| s.to_string())
+        else {
+            continue;
+        };
         let device = IosDevice::new(
             device["deviceProperties"]["name"]
                 .as_str()